--- app/app-routing.module.ts
@@ -1,35 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { DocsComponent } from './pages/docs/docs.component';
-import { SupportComponent } from './pages/support/support.component';
-import { Link1Component } from './pages/link1/link1.component';
-
-const routes: Routes = [
-	{
-		path: '',
-		loadChildren: () => import('./starter-home/starter-home.module').then(m => m.StarterHomeModule),
-		pathMatch: 'full'
-	},
-	{
-		path: 'catalog',
-		loadChildren: () => import('./pages/catalog/catalog.module').then(m => m.CatalogModule)
-	},
-	{
-		path: 'docs',
-		component: DocsComponent
-	},
-	{
-		path: 'support',
-		component: SupportComponent
-	},
-	{
-		path: 'link1',
-		component: Link1Component
-	}
-];
-
-@NgModule({
-	imports: [RouterModule.forRoot(routes, {})],
-	exports: [RouterModule]
-})
-export class AppRoutingModule { }
