--- app/app.component.html
@@ -1,4 +1,0 @@
-<app-header></app-header>
-<main class="cds--content">
-	<router-outlet></router-outlet>
-</main>
