--- app/app.component.scss
@@ -1,1 +1,0 @@
-@import "../assets/scss/scss-utils.scss";
