--- app/app.component.spec.ts
@@ -1,27 +1,0 @@
-import { TestBed, waitForAsync } from '@angular/core/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-
-import { UIShellModule } from 'carbon-components-angular';
-import { AppComponent } from './app.component';
-import { HeaderComponent } from './header/header.component';
-
-describe('AppComponent', () => {
-	beforeEach(waitForAsync(() => {
-		TestBed.configureTestingModule({
-			declarations: [
-				AppComponent,
-				HeaderComponent
-			],
-			imports: [
-				RouterTestingModule,
-				UIShellModule
-			]
-		}).compileComponents();
-	}));
-
-	it('should create the app', waitForAsync(() => {
-		const fixture = TestBed.createComponent(AppComponent);
-		const app = fixture.debugElement.componentInstance;
-		expect(app).toBeTruthy();
-	}));
-});
