--- app/header/header.component.html
@@ -1,22 +1,0 @@
-<div cdsTheme="g100">
-	<cds-header name="[Platform]" [route]="['/']">
-		<cds-header-navigation>
-			<cds-header-item routerLink="catalog">Catalog</cds-header-item>
-			<cds-header-item routerLink="docs">Docs</cds-header-item>
-			<cds-header-item routerLink="support">Support</cds-header-item>
-			<cds-header-menu title="Manage">
-				<cds-header-item routerLink="link1">Link 1</cds-header-item>
-				<cds-header-item>Link 2</cds-header-item>
-				<cds-header-item>Link 3</cds-header-item>
-			</cds-header-menu>
-		</cds-header-navigation>
-		<cds-header-global>
-			<cds-header-action title="action">
-				<svg cdsIcon="notification" size="20"></svg>
-			</cds-header-action>
-			<cds-header-action title="action">
-				<svg cdsIcon="user--avatar" size="20"></svg>
-			</cds-header-action>
-		</cds-header-global>
-	</cds-header>
-</div>
