--- app/header/header.component.spec.ts
@@ -1,27 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { HeaderComponent } from './header.component';
-import { UIShellModule } from 'carbon-components-angular';
-
-describe('HeaderComponent', () => {
-	let component: HeaderComponent;
-	let fixture: ComponentFixture<HeaderComponent>;
-
-	beforeEach(waitForAsync(() => {
-		TestBed.configureTestingModule({
-			declarations: [ HeaderComponent ],
-			imports: [ UIShellModule ]
-		})
-		.compileComponents();
-	}));
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(HeaderComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it('should create', () => {
-		expect(component).toBeTruthy();
-	});
-});
