--- app/pages/catalog/catalog-routing.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { CatalogComponent } from './catalog.component';
-
-const routes: Routes = [{
-    path: '',
-    component: CatalogComponent
-}];
-
-@NgModule({
-    imports: [RouterModule.forChild(routes)],
-    exports: [RouterModule]
-})
-export class CatalogComponentModule { }
