--- app/pages/catalog/catalog.component.html
@@ -1,19 +1,0 @@
-<p>Catalog page</p>
-
-<div style="margin-top:10rem">
-    <p class="subtitle">Country selector: Web component</p>
-    <div class="dropdown">
-       <cds-custom-dropdown
-          [title]="'Web component'"
-            [label]="'Select country'"
-           >
-          <cds-custom-dropdown-item
-            *ngFor="let item of dropdownItems"
-            [value]="item.content"
-            [selected]="item.selected"
-          >
-            {{ item.content }}
-          </cds-custom-dropdown-item>
-           </cds-custom-dropdown>
-    </div>
- </div>
