--- app/pages/docs/docs.component.html
@@ -1,1 +1,0 @@
-<p>Docs page content goes here..</p>
