--- app/pages/docs/docs.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { DocsComponent } from './docs.component';
-
-describe('DocsComponent', () => {
-	let component: DocsComponent;
-	let fixture: ComponentFixture<DocsComponent>;
-
-	beforeEach(async(() => {
-		TestBed.configureTestingModule({
-			declarations: [ DocsComponent ]
-		})
-		.compileComponents();
-	}));
-
-	beforeEach(() => {
-		fixture = TestBed.createComponent(DocsComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it('should create', () => {
-		expect(component).toBeTruthy();
-	});
-});
