--- app/pages/docs/docs.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'app-docs',
-	templateUrl: './docs.component.html',
-	styleUrls: ['./docs.component.scss']
-})
-export class DocsComponent implements OnInit {
-
-	constructor() { }
-
-	ngOnInit() {
-	}
-
-}
