--- app/pages/link1/link1.component.html
@@ -1,1 +1,0 @@
-<p>Link 1 page content goes here..</p>
