--- app/pages/link1/link1.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'app-link1',
-	templateUrl: './link1.component.html',
-	styleUrls: ['./link1.component.scss']
-})
-export class Link1Component implements OnInit {
-
-	constructor() { }
-
-	ngOnInit() {
-	}
-
-}
