--- app/pages/support/support.component.html
@@ -1,1 +1,0 @@
-<p>Support page content goes here..</p>
