--- app/pages/support/support.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-	selector: 'app-support',
-	templateUrl: './support.component.html',
-	styleUrls: ['./support.component.scss']
-})
-export class SupportComponent implements OnInit {
-
-	constructor() { }
-
-	ngOnInit() {
-	}
-
-}
