--- app/starter-home/starter-home/starter-home.component.scss
@@ -1,78 +1,0 @@
-@import "@carbon/layout";
-@import "@carbon/type";
-
-.starter-home {
-	.title {
-		@include type-style("productive-heading-05");
-	}
-
-	.subtitle {
-		@include type-style("productive-heading-02");
-		margin-bottom: 1rem;
-	}
-
-	.body-text {
-		margin-top: 1rem;
-		margin-bottom: 2rem;
-	}
-
-	.list {
-		max-width: 31.5rem;
-		padding-left: 3rem;
-		column-count: 2;
-
-		@media only screen and (max-width: 1055px) {
-			margin-top: 1rem;
-		}
-	}
-
-	.starter-tab-description {
-		margin-bottom: 1rem;
-	}
-
-	.theme-picker {
-		margin: 1rem 0;
-	}
-
-	.dropdown-container {
-		.dropdown {
-			display: flex;
-			justify-content: space-between;
-			gap: 1rem;
-
-			.cds--dropdown__wrapper {
-				width: 100%;
-			}
-		}
-	}
-
-	cds-tabs {
-		display: block;
-		margin-top: 3rem;
-	}
-
-	.builder-heading {
-		@include type-style("productive-heading-04");
-		margin-bottom: 1rem;
-	}
-
-	.builder-points {
-		margin-bottom: 2rem;
-		padding-left: 1rem;
-	}
-
-	.builder-image {
-		width: inherit;
-	}
-
-	ul {
-		list-style-type: disc;
-
-		li::before {
-			content: ''
-		}
-
-		line-height: 1.75rem;
-		font-size: 1.25rem;
-	}
-}
