--- assets/scss/base/_variables.scss
@@ -1,1 +1,0 @@
-
