--- assets/scss/carbon-override.scss
@@ -1,4 +1,0 @@
-// forces the navigation items to be displayed at all viewport sizes
-.cds--header__nav {
-	display: block;
-}
