--- assets/scss/scss-utils.scss
@@ -1,1 +1,0 @@
-@import "base/variables";
