--- environments/environment.prod.ts
@@ -1,3 +1,0 @@
-export const environment = {
-	production: true
-};
