+++ src/app/starter-home/starter-home/starter-home.component.ts
@@ -3,7 +3,8 @@ import { Component, OnInit, Renderer2, inject } from "@angular/core";
 
 import '@carbon/ibm-products-web-components/es/components/tearsheet/index.js';
 // The following are used for slotted fields
-import '@carbon/web-components/es/components/button/index.js';
+import '@carbon/web-components/es-custom/components/button/index.js';
+import '@carbon/web-components/es-custom/components/modal/index.js';
 
 @Component({
   selector: "app-starter-home",
@@ -134,11 +135,25 @@ export class StarterHomeComponent implements OnInit {
       selected: false,
     },
   ];
-  openTearsheet = false;
-  openModal = false;
+  // openTearsheet = false;
+  // openModal = false;
 
   private document = inject(DOCUMENT);
   private renderer = inject(Renderer2);
+
+  openWebComponentsModal() {
+    const modal = this.document.getElementById("web-components-modal");
+    if (modal) {
+      modal.setAttribute("open", "true");
+    }
+  }
+
+  openTearsheet() {
+    const modal = this.document.getElementById("tearsheet");
+    if (modal) {
+      modal.setAttribute("open", "true");
+    }
+  }
 
   ngOnInit() {
     // Apply default class to body
